+++ src/vm/mod.rs
@@ -581,6 +581,8 @@ pub struct Vm {
     pub(crate) realm_temporal_instant_prototypes: HashMap<usize, Value>,
     pub(crate) realm_temporal_duration_constructors: HashMap<usize, Value>,
     pub(crate) realm_temporal_duration_prototypes: HashMap<usize, Value>,
+    pub(crate) realm_temporal_plain_date_constructors: HashMap<usize, Value>,
+    pub(crate) realm_temporal_plain_date_prototypes: HashMap<usize, Value>,
     pub(crate) realm_temporal_plain_date_time_constructors: HashMap<usize, Value>,
     pub(crate) realm_temporal_plain_date_time_prototypes: HashMap<usize, Value>,
     pub(crate) realm_temporal_zoned_date_time_constructors: HashMap<usize, Value>,
@@ -1349,6 +1351,8 @@ impl Vm {
             realm_temporal_instant_prototypes: HashMap::new(),
             realm_temporal_duration_constructors: HashMap::new(),
             realm_temporal_duration_prototypes: HashMap::new(),
+            realm_temporal_plain_date_constructors: HashMap::new(),
+            realm_temporal_plain_date_prototypes: HashMap::new(),
             realm_temporal_plain_date_time_constructors: HashMap::new(),
             realm_temporal_plain_date_time_prototypes: HashMap::new(),
             realm_temporal_zoned_date_time_constructors: HashMap::new(),
@@ -3063,6 +3067,8 @@ impl Vm {
         self.realm_temporal_instant_prototypes.remove(&realm);
         self.realm_temporal_duration_constructors.remove(&realm);
         self.realm_temporal_duration_prototypes.remove(&realm);
+        self.realm_temporal_plain_date_constructors.remove(&realm);
+        self.realm_temporal_plain_date_prototypes.remove(&realm);
         self.realm_temporal_plain_date_time_constructors
             .remove(&realm);
         self.realm_temporal_plain_date_time_prototypes
@@ -3178,6 +3184,8 @@ impl Vm {
         push_realm_value!(realm_temporal_instant_prototypes);
         push_realm_value!(realm_temporal_duration_constructors);
         push_realm_value!(realm_temporal_duration_prototypes);
+        push_realm_value!(realm_temporal_plain_date_constructors);
+        push_realm_value!(realm_temporal_plain_date_prototypes);
         push_realm_value!(realm_temporal_plain_date_time_constructors);
         push_realm_value!(realm_temporal_plain_date_time_prototypes);
         push_realm_value!(realm_temporal_zoned_date_time_constructors);